fix: make TOS agreement text one semantic element
