+++ lib/routes/home/login/login_options_view.dart
@@ -1,12 +1,10 @@
-import 'package:flutter/gestures.dart';
 import 'package:flutter/material.dart';
 
 import 'package:go_router/go_router.dart';
-import 'package:url_launcher/url_launcher_string.dart';
 
-import 'package:fluffychat/config/app_config.dart';
 import 'package:fluffychat/l10n/l10n.dart';
 import 'package:fluffychat/routes/home/login/login.dart';
+import 'package:fluffychat/routes/home/login/tos_indicator.dart';
 import 'package:fluffychat/routes/home/p_sso_button.dart';
 import 'package:fluffychat/routes/home/pangea_logo_svg.dart';
 import 'package:fluffychat/routes/home/sso_provider_enum.dart';
@@ -124,34 +122,7 @@ class LoginOptionsViewState extends State<LoginOptionsView> {
                 ),
                 Padding(
                   padding: const EdgeInsets.all(12.0),
-                  child: RichText(
-                    textAlign: TextAlign.justify,
-                    text: TextSpan(
-                      text: L10n.of(context).byUsingPangeaChat,
-                      children: [
-                        TextSpan(
-                          text: L10n.of(context).termsAndConditions,
-                          style: TextStyle(
-                            decoration: TextDecoration.underline,
-                            color: theme.colorScheme.primary,
-                          ),
-                          recognizer: TapGestureRecognizer()
-                            ..onTap = () {
-                              launchUrlString(AppConfig.termsOfServiceUrl);
-                            },
-                        ),
-                        TextSpan(
-                          text: L10n.of(
-                            context,
-                          ).andCertifyIAmAtLeast13YearsOfAge,
-                        ),
-                      ],
-                      style: TextStyle(
-                        fontSize: 12,
-                        color: Theme.of(context).colorScheme.onSurface,
-                      ),
-                    ),
-                  ),
+                  child: TOSIndicator(),
                 ),
               ],
             ),
